Got JWT authentication strategy working
This code change uses the jwtStrategy by itself to go through the
process of logging into a site using username & password, generating
the jwt, returning it to the client, and the client using that just
created token to log into to the '/account' resource successfully.

Specific file changes are outlined below, but I wanted to highlight some
overall changes to the application in order to understand. You can either
integrate these changes as is, or restructure them to fit your intended
design.

CAVEAT: While the login to the account page is successful, the page
does not currently redirect to the account page. I believe that the redirect
can be accomplished by having the callback success function do a redirect
to a particular page and then have the page use the token to request
the profile information and update upon success.

END CAVEAT:

- auth/index.js - Removed basicStrategy

- auth/router.js - moved createAuthToken from auth/router.js to server.js

  removed login function

- auth/strategies.js - removed basicStrategy

- package-lock.json - Updated package

- package.json - Updated package

- public/account.js - changed LOGIN_URL & PROFILE_URL

- public/login.html - Changed name attributes of input boxes

- server.js - Added createAuthToken, removed basicStrategy, added
  "/login" method that finds user and creates JWT